web_admin/requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
cryptography==41.0.7
Werkzeug==2.3.7
Jinja2==3.1.2
python-dateutil==2.8.2 